mf_portfolio_manager/pyproject.toml
[project]
name = "mf_portfolio_manager"
version = "0.1.0"
description = "mf-portfolio-manager using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.165.1,<1.0.0"
]

[project.scripts]
mf_portfolio_manager = "mf_portfolio_manager.main:run"
run_crew = "mf_portfolio_manager.main:run"
train = "mf_portfolio_manager.main:train"
replay = "mf_portfolio_manager.main:replay"
test = "mf_portfolio_manager.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"

[dependency-groups]
dev = [
    "crewai>=0.203.1",
    "deprecated>=1.3.1",
    "ipykernel>=7.1.0",
    "langchain>=1.0.8",
    "langchain-core>=1.0.7",
    "langchain-openai>=1.0.3",
    "matplotlib>=3.10.7",
    "mftool>=3.0",
    "numpy>=2.2.6",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "yfinance>=0.2.66",
]
